Moved sections to src/components/common/ContactDetails
