+++ plugin-hrm-form/src/___tests__/components/case/ViewContact.test.js
@@ -9,7 +9,7 @@ import '../../mockGetConfig';
 
 import { UnconnectedViewContact } from '../../../components/case/ViewContact';
 import HrmTheme from '../../../styles/HrmTheme';
-import { ContactDetailsSections } from '../../../states/search/types';
+import { ContactDetailsSections } from '../../../components/common/ContactDetails';
 import { adaptFormToContactDetails } from '../../../components/case/ContactDetailsAdapter';
 
 expect.extend(toHaveNoViolations);
